+++ resell-shop-api/src/main/kotlin/com/hyuuny/resellshop/products/service/ProductCommand.kt
@@ -1,7 +1,6 @@
 package com.hyuuny.resellshop.products.service
 
 import com.hyuuny.resellshop.products.domain.Brand
-import java.math.BigDecimal
 import java.time.LocalDate
 
 data class CreateProductCommand(
@@ -9,7 +8,7 @@ data class CreateProductCommand(
     val nameEn: String,
     val brand: Brand,
     val nameKo: String,
-    val releasePrice: BigDecimal?,
+    val releasePrice: Long?,
     val modelNumber: String,
     val releaseDate: LocalDate?,
     val option: String,
